ticket-tracker/requirements.txt
streamlit
plotly
httpx
python-dotenv
pandas
